move extract_python_blocks to postprocess
