+++ utils/post_process.py
@@ -1,0 +1,7 @@
+import re
+
+def extract_python_blocks(text):
+    pattern = r"```python(.*?)```"
+    matches = re.findall(pattern, text, re.DOTALL)
+    combined = '\n'.join(match.strip() for match in matches)
+    return combined
